requirements.txt
alembic==1.7.5
algorithm==1.2.0
aniso8601==9.0.1
antiorm==1.2.1
black==21.12b0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
db==0.1.1
dependency-injection==1.2.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
lark-parser==0.12.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
password-strength==0.0.3.post2
pathspec==0.9.0
permission==0.4.1
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.27.0
required==0.4.0
six==1.16.0
SQLAlchemy==1.4.29
stripe==2.64.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
